JL8-20 reafactor : 플레이 리스트 쿼리 문 커스텀 훅으로 분리
